fix : 생성 후 바로 삭제시 에러 처리

+++ frontend/src/api/user.js
@@ -95,7 +95,9 @@ export const patchApi = async (data, access_token) => {
 };
 
 export const deleteApi = async (type, data, access_token) => {
-  console.log(typeof type);
+  if (data.search("create") === 0) {
+    return;
+  }
   try {
     const response = await axios.delete(
       `${API_BASE_URL}/user-info/${type}/${data}`,
